Moved benchparse to minibench.parse
